Support branches without upstream and show their status.
The current HEAD branch is now marked with an asterisk.

If the branch has no upstream, exceptions where catched and the message
"no upstream" is shown.

The stat of modifications is only shown for the current HEAD branch,
even if this branch has no upstream (support for local feature branches
or something similar).